chore: workflow improvement (#1989)
* Update verify.yml with setup-common-steps

This commit updates the verify.yml file to include the setup-common-steps job. The setup-common-steps job is responsible for setting up the common steps required for the verify workflow. By adding this job, we ensure that the necessary setup is performed before running the verify job.

The update includes the following changes:

- Added the setup-common-steps job to the jobs section.
- Added the needs keyword to the verify, browser-tests, node-tests, and node-tests-windows jobs, specifying that they depend on the setup-common-steps job.

* Update actions/setup-node and actions/checkout versions

This commit updates the versions of actions/setup-node and actions/checkout in the release.yml file.

* Rename 'setup-common-steps' to 'init' in verify.yml

Co-authored-by: gerjanvangeest <gerjanvangeest@users.noreply.github.com>

* Update setup-node and action/checkout to v3

* Update setup-node and action/checkout to v3 and rename setup-common-steps to init

---------

Co-authored-by: gerjanvangeest <gerjanvangeest@users.noreply.github.com>