fix(form-core): don't remove event listeners if _focusableNode was not set
This fix is similar to the previous commit.

Under very specific circumstances on a very complicated application logic, a file input field (that
extends from FocusMixin) gets created and immediately destroyed. This unfortunately causes an
error, because the `disconnectedCallback()` gets called to remove the event listeners for a
component that hasn't been fully set up.

> Uncaught TypeError: this._focusableNode is null